CameraSurfaceDrawingThread の引数をCameraWrapper からCamera2Recorder に変更 ソースコードを整理

+++ dConnectDevicePlugin/dConnectDeviceHost/app/src/main/java/org/deviceconnect/android/deviceplugin/host/recorder/camera/Camera2PreviewServer.java
@@ -39,7 +39,7 @@ abstract class Camera2PreviewServer extends AbstractPreviewServer {
         videoQuality.setVideoWidth(settings.getPreviewSize().getWidth());
         videoQuality.setVideoHeight(settings.getPreviewSize().getHeight());
         videoQuality.setBitRate(settings.getPreviewBitRate());
-        videoQuality.setFrameRate(settings.getPreviewMaxFrameRate());
+        videoQuality.setFrameRate((int) settings.getPreviewMaxFrameRate());
         videoQuality.setIFrameInterval(settings.getPreviewKeyFrameInterval());
     }
